requirements.txt
cssselect==1.2.0
lxml==5.3.0
markdown-it-py==3.0.0
mdurl==0.1.2
packaging==24.1
types-lxml==2024.8.7
typing_extensions==4.12.2
